agent-pipecat/requirements.txt
pipecat-ai[livekit,google,deepgram,elevenlabs,silero,simli]>=0.0.100,<0.1.0
pipecat-ai-flows>=0.0.5,<0.1.0
python-dotenv>=1.0.0,<2.0.0
aiohttp>=3.9.0,<4.0.0
openai>=2.0.0,<3.0.0
